requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
PyMuPDF==1.25.5
fpdf2==2.8.3
cohere==5.15.0
pydantic-settings==2.9.1
python-dotenv==1.1.0